+++ HISTORY.rst
@@ -237,6 +237,7 @@
 **Enhancements**
 
 - 2471_: use Vulture CLI tool to detect dead code.
+- 2385_, [Linux]: add support for zfs arc memory stats. (patch by Hudson Gerwing)
 
 **Bug fixes**